Blatantly copy any built craftbukkit and spigot jars.
Even if you specify what version to build, if BuildTools
can't find it, it will build the latest. And in which
case, the copy process will fail.

We still want to preserve the compiled jar even if it's
the wrong version. That way, we can simply recreate the
container, passing the correct MINECRAFT_VERSION variable.