Added ONBUILD trigger to Spigot for custom config packaging.
The main problem with maintaining a minecraft server
is the separation of data and configuration.

Docker images usually come with the configuration, and
the data are stored in volumes. This won't be an issue
if Docker provides data volume layering. However based on
my research, that feature is relatively new.

To support the need of packaging configuration separately
while maintaining the use data volumes, I use rsync to
copy the updated files to the MINECRAFT_HOME folder. This
is done on a container's first start. Subsequent restarts
of the container will not recopy the files. This ensures
that any changes to the configuration at runtime will not
be accidentally overridden during a server restart.